faster file copy

+++ src/snabel/type.hpp
@@ -28,7 +28,7 @@ namespace snabel {
     opt<func<bool (Scope &, const Box &, bool)>> call;
     opt<func<Iter::Ref (const Box &)>> iter;
     opt<func<bool (const Box &, IOBuf &)>> read;
-    opt<func<int64_t (const Box &, const unsigned char *, size_t)>> write;
+    opt<func<int64_t (const Box &, const unsigned char *, int64_t)>> write;
     
     Type(const str &n);
     Type(const Type &) = delete;
